chore: remove old constructors with double

+++ CurrencyConverter/src/main/java/ar/edu/itba/dps/exchange/main/Main.java
@@ -7,6 +7,7 @@ import ar.edu.itba.dps.exchange.infrastructure.http.UnirestHttpClient;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.math.BigDecimal;
 import java.time.Clock;
 import java.util.Currency;
 
@@ -19,7 +20,7 @@ public final class Main {
 			final var httpClient = new UnirestHttpClient();
 			final var provider = new FreeCurrencyApiProvider(httpClient);
 			final var converter = new CurrencyConverter(provider, Clock.systemUTC());
-			LOG.info("{}", converter.convert(new Money(Currency.getInstance("EUR"), 100), Currency.getInstance("USD")));
+			LOG.info("{}", converter.convert(new Money(Currency.getInstance("EUR"), BigDecimal.valueOf(100)), Currency.getInstance("USD")));
 		} catch (final RuntimeException e) {
 			LOG.error("Conversion demo failed", e);
 			System.exit(1);
